requirements.txt
Flask==2.3.3
pymongo==4.5.0
pandas==2.1.1
google-generativeai==0.3.0
python-dotenv==1.0.0
gunicorn